lg-multi-agent-supervisor/pyproject.toml
[project]
name = "lg-multi-agent-supervisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.6",
    "langchain-community>=0.3.4",
    "langchain-openai>=0.2.4",
    "langgraph>=0.2.42",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.0",
    "langchain-experimental>=0.3.2",
]
